Split code a bit
